requirements.txt
tiktoken
openai
chromadb
langchain

beautifulsoup4
nltk

pytest
pytest-asyncio
discord-ext-test

discord.py
py-cord

scrapy
scrapy-fake-useragent